Add search for sites of group, fix add/delete sites to/from group

+++ src/main/java/kz/sitehealthtrackerbackend/site_health_tracker_backend/model/Site.java
@@ -31,7 +31,7 @@ public class Site extends BaseEntity<Long> {
     @Column(name = "url", unique = true)
     private String url;
     @Column(name = "site_health_check_interval")
-    private Integer siteHealthCheckInterval;
+    private Long siteHealthCheckInterval;
     @Enumerated(EnumType.STRING)
     @Column(name = "status", columnDefinition = "site_status")
     @Type(PostgreSQLEnumType.class)
